simple_requirements.txt
# Updated requirements for Python 3.13
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
streamlit==1.28.0
numpy>=2.0.0,<2.3.0  # Ensure we get a stable, available version
scikit-learn==1.4.2  # Updated version with Python 3.13 support
certifi==2024.2.2
